#59 alarm 로직 분리
