+++ app/src/main/java/com/thequietz/travelog/menu/MenuFragment.kt
@@ -1,12 +1,6 @@
 package com.thequietz.travelog.menu
 
-import android.app.AlarmManager
-import android.app.PendingIntent
-import android.content.Context
-import android.content.Intent
 import android.os.Bundle
-import android.os.SystemClock
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -18,10 +12,10 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import com.thequietz.travelog.R
 import com.thequietz.travelog.databinding.FragmentMenuBinding
-import com.thequietz.travelog.menu.alarm.AlarmReceiver
+import com.thequietz.travelog.menu.alarm.AlarmType
+import com.thequietz.travelog.menu.alarm.cancelAlarm
+import com.thequietz.travelog.menu.alarm.registerAlarm
 import dagger.hilt.android.AndroidEntryPoint
-import java.text.SimpleDateFormat
-import java.util.Calendar
 
 @AndroidEntryPoint
 class MenuFragment : Fragment() {
@@ -85,7 +79,6 @@ class MenuFragment : Fragment() {
         binding.spinnerRecord.onItemSelectedListener = object : AdapterView.OnItemSelectedListener {
             override fun onItemSelected(p0: AdapterView<*>?, p1: View?, position: Int, p3: Long) {
                 viewModel.recordTimeChange(position)
-
             }
 
             override fun onNothingSelected(p0: AdapterView<*>?) {
@@ -94,39 +87,14 @@ class MenuFragment : Fragment() {
     }
 
     fun initAlarm() {
-        val context = requireActivity().applicationContext
-        val alarmManager = requireActivity().getSystemService(Context.ALARM_SERVICE) as AlarmManager
-
-        val intent = Intent(context, AlarmReceiver::class.java)
-        intent.putExtra("name", "무슨무슨여행~^^@/")
-        val pendingIntent = PendingIntent.getBroadcast(
-            context,
-            AlarmReceiver.NOTIFICATION_ID,
-            intent,
-            PendingIntent.FLAG_UPDATE_CURRENT
-        )
-
-        val calendar = Calendar.getInstance().apply {
-            val date = "2021-11-17 10:47:00"
-            val dateFormat = SimpleDateFormat("yyyy-MM-dd HH:mm:ss")
-            time = dateFormat.parse(date)
-        }
-
         binding.btnAlarm.setOnCheckedChangeListener(
             CompoundButton.OnCheckedChangeListener { _, isChecked ->
-                val message = if (isChecked) {
-                    val triggerTime = (SystemClock.elapsedRealtime() + 10 * 1000)
-                    alarmManager.set(
-                        AlarmManager.RTC,
-                        calendar.timeInMillis,
-                        pendingIntent
-                    )
-                    "Alarm On"
-                } else {
-                    alarmManager.cancel(pendingIntent)
-                    "Alarm off"
+                when (isChecked) {
+                    true -> registerAlarm(requireContext(),
+                        AlarmType.Schedule,
+                        null)
+                    else -> cancelAlarm()
                 }
-                Log.e("Alarm", message)
             }
         )
     }
